+++ Source/EventFlow.MsSql/IMsSqlDatabaseMigrator.cs
@@ -21,23 +21,12 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // 
-using System;
-using System.Collections.Generic;
-using System.Linq;
 
-namespace EventFlow.MsSql.Exceptions
+using EventFlow.Sql.Migrations;
+
+namespace EventFlow.MsSql
 {
-    public class MssqlMigrationException : Exception
+    public interface IMsSqlDatabaseMigrator : ISqlDatabaseMigrator
     {
-        public IReadOnlyCollection<string> Scripts { get; private set; }
-
-        public MssqlMigrationException(
-            IEnumerable<string> scripts, 
-            string message,
-            Exception innerException)
-            : base(message, innerException)
-        {
-            Scripts = scripts.ToList();
-        }
     }
 }