+++ Source/EventFlow.Sql/Connections/ISqlConfiguration.cs
@@ -22,26 +22,10 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-using System;
-using EventFlow.Core;
-using EventFlow.Sql.Connections;
-
-namespace EventFlow.MsSql
+namespace EventFlow.Sql.Connections
 {
-    public class MsSqlConfiguration : SqlConfiguration<IMsSqlConfiguration>, IMsSqlConfiguration
+    public interface ISqlConfiguration
     {
-        public static MsSqlConfiguration New => new MsSqlConfiguration();
-
-        public RetryDelay TransientRetryDelay { get; private set; } = RetryDelay.Between(
-            TimeSpan.FromMilliseconds(50),
-            TimeSpan.FromMilliseconds(100));
-
-        private MsSqlConfiguration() { }
-
-        public IMsSqlConfiguration SetTransientRetryDelay(RetryDelay retryDelay)
-        {
-            TransientRetryDelay = retryDelay;
-            return this;
-        }
+        string ConnectionString { get; }
     }
 }